+++ crates/zng-wgt-rule-line/src/lib.rs
@@ -12,6 +12,12 @@
 use zng_wgt::{margin, prelude::*};
 use zng_wgt_access::{AccessRole, access_role};
 
+pub mod hr;
+pub mod vr;
+
+mod collapse;
+pub use collapse::*;
+
 /// Draws a horizontal or vertical rule line.
 #[widget($crate::RuleLine)]
 pub struct RuleLine(WidgetBase);
@@ -45,7 +51,7 @@ pub fn stroke_thickness(thickness: impl IntoVar<Length>) {}
 
 /// Line length.
 ///
-/// Set to [`Default`] to fill available length without requesting any length.
+/// Set to [`Default`] to fill available length.
 ///
 /// [`Default`]: Length::Default
 #[property(CONTEXT, capture, default(Length::Default), widget_impl(RuleLine))]
@@ -83,7 +89,15 @@ fn on_build(wgt: &mut WidgetBuilding) {
                 .sub_var_render(&color)
                 .sub_var_render(&style);
         }
+        UiNodeOp::Info { info } => {
+            info.flag_meta(*COLLAPSABLE_LINE_ID);
+        }
         UiNodeOp::Measure { desired_size, .. } => {
+            if COLLAPSE_SCOPE.collapse(WIDGET.id()) {
+                *desired_size = PxSize::zero();
+                return;
+            }
+
             let metrics = LAYOUT.metrics();
             let default_stroke = Dip::new(1).to_px(metrics.scale_factor());
 
@@ -98,7 +112,13 @@ fn on_build(wgt: &mut WidgetBuilding) {
                 ),
             };
         }
-        UiNodeOp::Layout { final_size, .. } => {
+        UiNodeOp::Layout { final_size, wl } => {
+            if COLLAPSE_SCOPE.collapse(WIDGET.id()) {
+                wl.collapse();
+                *final_size = PxSize::zero();
+                return;
+            }
+
             let metrics = LAYOUT.metrics();
             let default_stroke = Dip::new(1).to_px(metrics.scale_factor());
 
@@ -121,6 +141,9 @@ fn on_build(wgt: &mut WidgetBuilding) {
             *final_size = b;
         }
         UiNodeOp::Render { frame } => {
+            if bounds.is_empty() {
+                return;
+            }
             let bounds = PxRect::from_size(bounds);
             let orientation = orientation.get();
             let color = color.get();
@@ -129,130 +152,4 @@ fn on_build(wgt: &mut WidgetBuilding) {
         }
         _ => {}
     }));
-}
-
-/// Horizontal rule line.
-pub mod hr {
-    use zng_wgt::prelude::*;
-
-    /// Draws an horizontal [`RuleLine!`](struct@crate::RuleLine).
-    #[widget($crate::hr::Hr)]
-    pub struct Hr(super::RuleLine);
-    impl Hr {
-        fn widget_intrinsic(&mut self) {
-            widget_set! {
-                self;
-                orientation = LineOrientation::Horizontal;
-                color = COLOR_VAR;
-                stroke_thickness = STROKE_THICKNESS_VAR;
-                line_style = LINE_STYLE_VAR;
-                margin = MARGIN_VAR;
-            }
-        }
-    }
-
-    context_var! {
-        /// Line color, inherits from [`FONT_COLOR_VAR`].
-        ///
-        /// [`FONT_COLOR_VAR`]: zng_wgt_text::FONT_COLOR_VAR
-        pub static COLOR_VAR: Rgba = zng_wgt_text::FONT_COLOR_VAR.map(|c| c.with_alpha(30.pct()));
-
-        /// Line stroke thickness, default is `1.dip()`
-        pub static STROKE_THICKNESS_VAR: Length = 1.dip();
-
-        /// Line style, default is `Solid`.
-        pub static LINE_STYLE_VAR: LineStyle = LineStyle::Solid;
-
-        /// Margin around line.
-        ///
-        /// Is `(4, 0)` by default, 4 top-bottom, 0 left-right.
-        pub static MARGIN_VAR: SideOffsets = (4, 0);
-    }
-
-    /// Sets the [`COLOR_VAR`] that affects all horizontal rules inside the widget.
-    #[property(CONTEXT, default(COLOR_VAR))]
-    pub fn color(child: impl IntoUiNode, color: impl IntoVar<Rgba>) -> UiNode {
-        with_context_var(child, COLOR_VAR, color)
-    }
-
-    /// Sets the [`STROKE_THICKNESS_VAR`] that affects all horizontal rules inside the widget.
-    #[property(CONTEXT, default(STROKE_THICKNESS_VAR))]
-    pub fn stroke_thickness(child: impl IntoUiNode, thickness: impl IntoVar<Length>) -> UiNode {
-        with_context_var(child, STROKE_THICKNESS_VAR, thickness)
-    }
-
-    /// Sets the [`LINE_STYLE_VAR`] that affects all horizontal rules inside the widget.
-    #[property(CONTEXT, default(LINE_STYLE_VAR))]
-    pub fn line_style(child: impl IntoUiNode, style: impl IntoVar<LineStyle>) -> UiNode {
-        with_context_var(child, LINE_STYLE_VAR, style)
-    }
-
-    /// Sets the [`MARGIN_VAR`] that affects all horizontal rules inside the widget.
-    #[property(CONTEXT, default(MARGIN_VAR))]
-    pub fn margin(child: impl IntoUiNode, margin: impl IntoVar<SideOffsets>) -> UiNode {
-        with_context_var(child, MARGIN_VAR, margin)
-    }
-}
-
-/// Vertical rule line.
-pub mod vr {
-    use zng_wgt::prelude::*;
-
-    /// Draws a vertical [`RuleLine!`](struct@crate::RuleLine).
-    #[widget($crate::vr::Vr)]
-    pub struct Vr(super::RuleLine);
-    impl Vr {
-        fn widget_intrinsic(&mut self) {
-            widget_set! {
-                self;
-                orientation = LineOrientation::Vertical;
-                color = COLOR_VAR;
-                stroke_thickness = STROKE_THICKNESS_VAR;
-                line_style = LINE_STYLE_VAR;
-                margin = MARGIN_VAR;
-            }
-        }
-    }
-
-    context_var! {
-        /// Line color, inherits from [`FONT_COLOR_VAR`].
-        ///
-        /// [`FONT_COLOR_VAR`]: zng_wgt_text::FONT_COLOR_VAR
-        pub static COLOR_VAR: Rgba = zng_wgt_text::FONT_COLOR_VAR.map(|c| c.with_alpha(30.pct()));
-
-        /// Line stroke thickness, default is `1.dip()`
-        pub static STROKE_THICKNESS_VAR: Length = 1.dip();
-
-        /// Line style, default is `Solid`.
-        pub static LINE_STYLE_VAR: LineStyle = LineStyle::Solid;
-
-        /// Margin around line.
-        ///
-        /// Is `(0, 4)` by default, 0 top-bottom, 4 left-right.
-        pub static MARGIN_VAR: SideOffsets = (0, 4);
-    }
-
-    /// Sets the [`COLOR_VAR`] that affects all vertical rules inside the widget.
-    #[property(CONTEXT, default(COLOR_VAR))]
-    pub fn color(child: impl IntoUiNode, color: impl IntoVar<Rgba>) -> UiNode {
-        with_context_var(child, COLOR_VAR, color)
-    }
-
-    /// Sets the [`STROKE_THICKNESS_VAR`] that affects all vertical rules inside the widget.
-    #[property(CONTEXT, default(STROKE_THICKNESS_VAR))]
-    pub fn stroke_thickness(child: impl IntoUiNode, thickness: impl IntoVar<Length>) -> UiNode {
-        with_context_var(child, STROKE_THICKNESS_VAR, thickness)
-    }
-
-    /// Sets the [`LINE_STYLE_VAR`] that affects all vertical rules inside the widget.
-    #[property(CONTEXT, default(LINE_STYLE_VAR))]
-    pub fn line_style(child: impl IntoUiNode, style: impl IntoVar<LineStyle>) -> UiNode {
-        with_context_var(child, LINE_STYLE_VAR, style)
-    }
-
-    /// Sets the [`MARGIN_VAR`] that affects all vertical rules inside the widget.
-    #[property(CONTEXT, default(MARGIN_VAR))]
-    pub fn margin(child: impl IntoUiNode, margin: impl IntoVar<SideOffsets>) -> UiNode {
-        with_context_var(child, MARGIN_VAR, margin)
-    }
 }